requirements.txt
django
virtualenv
django-rest-framework
django-cors-headers
djangorestframework-jwt